moved update inner bounds test
